requirements.txt
Beaker==1.6.2
BeautifulSoup==3.2.0
Chameleon==2.7.0
FormEncode==1.2.4
Mako==0.5.0
MarkupSafe==0.15
Paste==1.7.5.1
PasteDeploy==1.5.0
PasteScript==1.7.5
PyYAML==3.10
Pygments==1.4
Routes==1.12.3
SQLAlchemy==0.7.3
WebOb==1.2b2
WebTest==1.3.3
alembic==0.3.0
certifi==0.0.7
chardet==1.0.1
cryptacular==1.4
httplib2==0.7.2
mock==0.7.2
nose==1.1.2
oauth2==1.5.211
pbkdf2==1.3
psycopg2==2.4.5
pyramid==1.3a2
pyramid-beaker==0.6.1
pyramid-debugtoolbar==0.9.7
pyramid-tm==0.3
python-openid==2.2.5
repoze.lru==0.4
requests==0.10.1
transaction==1.2.0
translationstring==0.4
velruse==0.20a1
venusian==1.0a2
waitress==0.8.1
wsgiref==0.1.2
zope.deprecation==3.5.0
zope.interface==3.8.0
zope.sqlalchemy==0.7
